Changed redirection from landing page
