+++ cla-landing-page/src/app/components/page-not-found/page-not-found.component.ts
@@ -1,9 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import { AppSettings } from 'src/app/config/app-settings';
-import { REDIRECT_AUTH_ROUTE } from 'src/app/config/auth-utils';
-import { EnvConfig } from 'src/app/config/cla-env-utils';
-import { AuthService } from 'src/app/core/services/auth.service';
-import { StorageService } from 'src/app/core/services/storage.service';
 
 @Component({
   selector: 'app-page-not-found',
@@ -12,21 +7,9 @@ import { StorageService } from 'src/app/core/services/storage.service';
 })
 export class PageNotFoundComponent implements OnInit {
   message: string;
-  constructor(
-    private authService: AuthService,
-    private storageService: StorageService
-  ) { }
+  constructor() { }
 
   ngOnInit(): void {
-    if (this.authService.loggedIn) {
-      const type = JSON.parse(this.storageService.getItem('type'));
-      this.message = 'Wait we are redirecting you to ' + (type === 'Projects' ? 'Project' : 'Corporate') + ' Console.';
-      const redirectConsole = (type === 'Projects') ? AppSettings.PROJECT_CONSOLE_LINK : AppSettings.CORPORATE_CONSOLE_LINK;
-      setTimeout(() => {
-        window.open(EnvConfig.default[redirectConsole] + REDIRECT_AUTH_ROUTE, '_self');
-      }, 1500);
-    } else {
-      this.message = 'The page you are looking for was not found.';
-    }
+    this.message = 'The page you are looking for was not found.';
   }
 }
